# Documentation/api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/payment-accounts": {
      "post": {
        "operationId": "PaymentAccountsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment account created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentAccountResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "409": {
            "description": "Payment account name already exists",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create payment account",
        "tags": ["Payment Accounts"]
      },
      "get": {
        "operationId": "PaymentAccountsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["CASH", "BANK", "WALLET", "CARD"],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["ACTIVE", "INACTIVE", "ALL"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment account list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentAccountListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "List payment accounts",
        "tags": ["Payment Accounts"]
      }
    },
    "/api/v1/payment-accounts/{id}": {
      "get": {
        "operationId": "PaymentAccountsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Payment account UUID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment account detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentAccountResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Payment account not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get payment account by id",
        "tags": ["Payment Accounts"]
      },
      "patch": {
        "operationId": "PaymentAccountsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Payment account UUID",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment account updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentAccountResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Payment account not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "409": {
            "description": "Payment account name already exists",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Update payment account",
        "tags": ["Payment Accounts"]
      }
    },
    "/api/v1/payment-accounts/{id}/status": {
      "patch": {
        "operationId": "PaymentAccountsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Payment account UUID",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateStatusDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment account status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentAccountResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Payment account not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Update payment account status",
        "tags": ["Payment Accounts"]
      }
    },
    "/api/v1/payment-accounts/{id}/balance": {
      "get": {
        "operationId": "PaymentAccountsController_getBalance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Payment account UUID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment account balance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentAccountBalanceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Payment account not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get payment account balance",
        "tags": ["Payment Accounts"]
      }
    },
...
    "/api/v1/reports/payment-accounts/{id}/balance": {
      "get": {
        "operationId": "ReportsController_getPaymentAccountBalance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Payment account UUID",
            "schema": { "type": "string" }
          },
          {
            "name": "asOfDate",
            "required": false,
            "in": "query",
            "description": "Point-in-time date (defaults to today)",
            "schema": { "example": "2026-02-20", "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Payment account balance with breakdown" },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Payment account not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Payment account balance report (point-in-time)",
        "tags": ["Reports"]
      }
    },
...
  },
  "info": {
    "title": "Persona Finance API",
    "description": "Multi-tenant finance system. Standard error response format includes: `statusCode`, `message`, `errors[]`, `timestamp`, `path`, `requestId`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "ApiErrorField": {
        "type": "object",
        "properties": {
          "field": { "type": "string", "example": "name" },
          "message": { "type": "string", "example": "must not be empty" }
        },
        "required": ["field", "message"]
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "example": 400 },
          "message": { "type": "string", "example": "Validation failed" },
          "errors": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ApiErrorField" }
          },
          "timestamp": {
            "type": "string",
            "example": "2026-02-11T12:00:00.000Z"
          },
          "path": { "type": "string", "example": "/api/v1/customers" },
          "requestId": { "type": "string", "example": "req_123" }
        },
        "required": ["statusCode", "message", "timestamp", "path"]
      },
...
      "UpdateStatusDto": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE"] },
          "reason": { "type": "string", "example": "Duplicate record" }
        },
        "required": ["status"]
      },
...
      "CreatePaymentAccountDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "Main Cash" },
          "type": {
            "type": "string",
            "enum": ["CASH", "BANK", "WALLET", "CARD"]
          },
          "openingBalance": {
            "type": "number",
            "example": 25000,
            "default": 0,
            "description": "Opening balance in PKR (integer)"
          }
        },
        "required": ["name", "type"]
      },
      "PaymentAccountComputedDto": {
        "type": "object",
        "properties": {
          "currentBalance": {
            "type": "number",
            "example": 0,
            "description": "Derived current balance in PKR (integer)"
          },
          "totalIn": {
            "type": "number",
            "example": 0,
            "description": "Derived total in amount in PKR (integer)"
          },
          "totalOut": {
            "type": "number",
            "example": 0,
            "description": "Derived total out amount in PKR (integer)"
          },
          "lastTransactionDate": {
            "type": "object",
            "example": null,
            "nullable": true
          }
        },
        "required": [
          "currentBalance",
          "totalIn",
          "totalOut",
          "lastTransactionDate"
        ]
      },
      "PaymentAccountResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "9f4b6e2c-0a2d-4cc5-8c4d-1a4a88c81a88"
          },
          "tenantId": {
            "type": "string",
            "example": "6c6f7f48-3d5b-4a3f-9b1d-9c0d73b0c3d2"
          },
          "name": { "type": "string", "example": "Main Cash" },
          "type": { "type": "string", "example": "CASH" },
          "openingBalance": {
            "type": "number",
            "example": 25000,
            "description": "Opening balance in PKR (integer)"
          },
          "status": { "type": "string", "example": "ACTIVE" },
          "createdAt": {
            "type": "string",
            "example": "2026-02-11T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-02-11T10:00:00.000Z"
          },
          "createdBy": {
            "type": "object",
            "example": "d2f2c7b5-0c2a-4aa2-9c60-6b3f94b7e8d4"
          },
          "_computed": {
            "$ref": "#/components/schemas/PaymentAccountComputedDto"
          }
        },
        "required": [
          "id",
          "tenantId",
          "name",
          "type",
          "openingBalance",
          "status",
          "createdAt",
          "updatedAt",
          "_computed"
        ]
      },
      "PaymentAccountListMetaDto": {
        "type": "object",
        "properties": {
          "page": { "type": "number", "example": 1 },
          "limit": { "type": "number", "example": 20 },
          "total": { "type": "number", "example": 120 },
          "totalPages": { "type": "number", "example": 6 }
        },
        "required": ["page", "limit", "total", "totalPages"]
      },
      "PaymentAccountListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentAccountResponseDto"
            }
          },
          "meta": { "$ref": "#/components/schemas/PaymentAccountListMetaDto" }
        },
        "required": ["data", "meta"]
      },
      "UpdatePaymentAccountDto": {
        "type": "object",
        "properties": { "name": { "type": "string", "example": "Main Cash" } }
      },
      "PaymentAccountBalanceResponseDto": {
        "type": "object",
        "properties": {
          "paymentAccountId": {
            "type": "string",
            "example": "9f4b6e2c-0a2d-4cc5-8c4d-1a4a88c81a88"
          },
          "openingBalance": {
            "type": "number",
            "example": 25000,
            "description": "Opening balance in PKR (integer)"
          },
          "totalIn": {
            "type": "number",
            "example": 150000,
            "description": "Total in amount in PKR (integer)"
          },
          "totalOut": {
            "type": "number",
            "example": 50000,
            "description": "Total out amount in PKR (integer)"
          },
          "currentBalance": {
            "type": "number",
            "example": 125000,
            "description": "Current balance in PKR (integer)"
          }
        },
        "required": [
          "paymentAccountId",
          "openingBalance",
          "totalIn",
          "totalOut",
          "currentBalance"
        ]
      },
...
    }
  }
}